% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseShoppingCart.R
\name{analyseShoppingCart}
\alias{analyseShoppingCart}
\title{analyse the shopping cart}
\usage{
analyseShoppingCart(shoppingCart)
}
\arguments{
\item{shoppingCart}{shopping cart as an data.frame with one row for one item}
}
\value{
It returns a list with four numeric and named values.
}
\description{
This function analyses the content of a online shopping cart. It
looks for the following information: How many items are sold in total, how
many different items do we have, what's the maximum number of one item and
at least how often do we have this 'maximum number of one item' within the
shopping cart.
}
\examples{
```
books <- dplyr::tibble(
  itemID = 1:5,
  name = c(
    "Stein der Weisen",
    "Kammer des Schreckens",
    "Gefangene von Askaban",
    "Feuerkelch",
    "Orden des Phönix"
  )
)

set.seed(1)
randomNumbers<- sample(1:10, 5, replace = T)
shoppingCart <- dplyr::bind_cols(books, number = randomNumbers)
analyseShoppingCart(shoppingCart)
```
}
